PEDBP

Data and utilities for estimating pediatric blood pressure percentiles by sex, age, and optionally height (stature) as described in Martin et.al. (2022) doi:10.1001/jamanetworkopen.2022.36918 . Blood pressure percentiles for children under one year of age come from Gemelli et.al. (1990) doi:10.1007/BF02171556 . Estimates of blood pressure percentiles for children at least one year of age are informed by data from the National Heart, Lung, and Blood Institute (NHLBI) and the Centers for Disease Control and Prevention (CDC) doi:10.1542/peds.2009-2107C or from Lo et.al. (2013) doi:10.1542/peds.2012-1292 . The flowchart for selecting the informing data source comes from Martin et.al. (2022) doi:10.1542/hpeds.2021-005998 .

See also

Author

Maintainer: Peter DeWitt peter.dewitt@cuanschutz.edu (ORCID)

Other contributors:

Blake Martin blake.martin@childrenscolorado.org (ORCID) [contributor]

David Albers david.albers@cuanschutz.edu (ORCID) [contributor]

Tell Bennett tell.bennett@cuanschutz.edu (ORCID) [contributor]